Simplifying the configuration a bit.
We no longer need the additional macros in aclocal.m4, so that's been removed.

I removed the --with-swing option, but generalized it into --with-classpath

Here's a sample configuration that I use:

JAVAC=/usr/local/bin/jikes ./configure \
 --with-classpath=/usr/local/java/lib/classes.zip:/usr/local/swing/swing.jar \
 --with-devtarget

The --with-devtarget changes the default target to 'omcore', which just
builds the class files, not jars or documentation.

You could also use '--with-devtarget=myFavoriteTarget', to set your
own default target.


git-svn-id: svn+ssh://src.dsl.bbn.com/svn/openmap/trunk@156 f68c23cc-7f87-4fbc-acfa-d9760b90acc1