Added some data/vpf vmaplv0 configuration files with more instructions on how to use the VPFFeatureLayer.
Fixed the VPF attribute fetching with proper indexes.  The VPF package
api changed in some deep methods, to allow the proper feature id to be
passed to the place where it's needed to fetch 	attributes.

Added intersets method to DataBounds.  Can be tested with
new	BoundsTestLayer in layer.test package.