Refactored the LayerHandler so it uses a List<Layer> internally to manage Layers.
Created the LayerConfigurationListener interface and
LayerConfigurationListenerSupport object, both of which are used by the
LayerHandler to pre-notify components when the active layers for the
application are about to change.  The LayerConfigurationListener has the
opportunity to make modifications to the layer configuration, or change
visibility of other layers, etc.

Added the BasicLayerConfigurationListener, an example component that
changes the projection of the MapBean to CADRG if an RpfLayer is active.