MVTMapTileFactory added.
The VectorMapTileFactory was modified to move tile_data fetching to an
overrideable method.  The default implementation still reads from
mbtiles files.  The new MVTMapTileFactory uses a TilePathBuilder to go
to a server to fetch tile data, so the transfer is vector data.  The
default fileExt variable is now accessed through getters/setters.