Feature additions: TileGrabber, gap decoration and gapped projected shape.  Mods to DemoLayer and ColorFactory.
TileGrabber pulls tiles from one location and stores them to another.  Gaps are strokes that are two parallel lines on either side of a source line.  The GapDecoration attempts to do this with a stroke, the GAPPED OMShape addition is a different approach.  DemoLayer has tooltips added to decorated shapes so you can tell which is which, ColorFactory has slight bug fix for updating rgb to rrggbb.