Incorporated Gary B's changes to handle circles in the CSVLocationHandler, using indexes set in the properties describing the location file.
URLRasterLocation modified to use OMScalingIcons instead of OMRasters, added radius and radius units to CSVLocationHandler, which are used if they are defined and the iconIndex for an image URL isn't defined.  GraphicAttributes.setTo modified to not include RenderType when changing settings, should never have been used.  RenderType is only set if the OMGraphic render type is unknown.