LayerHandler and Layer GUI updates to property handle non-removable Layers.  Refactoring of LayerAddPanel and MenuList/MenuBar relationship.
Updated LayerHandler to manage decisions for non-removable layers properly.  Updated LayersPanel and LayersMenu to respect the non-removable status of layers, disabling controls that would let those layers be removed from the map.  MenuList updated to manage MenuBar and Menu, with MenuBar being less aggressive capturing JMenus found in MapHandler.  LayerAddPanel provides a text field for the user to enter classname of layers to add if it isn't initialized with a selection of layers as options.  MapTileServlet updated to allow setting of Leaflet javascript location, which can be local if internet isn't available.