Adjusted Planet constant value for meter circumference calc, added Proj.inverse(Shape) method.  Added JUnit test for geo conversions.
The Planet meter circumference constant for double was using a float constant in it, updated it to double for consistency.  The Projection class has an inverse method to inverse project a java.awt.Shape object, it handles iterating through all the points for you to create a new Shape object.